index int64

+++ models.py
@@ -79,7 +79,7 @@ class PNA(torch.nn.Module):
     ):
         x, edge_index, edge_attr, batch = (
             data.x.to(torch.float),
-            data.edge_index,
+            data.edge_index.to(torch.int64),
             data.edge_attr.to(torch.float),
             data.batch
         )
